chore: extract connection suite into src/suites/connection.ts
Part of REAL-1037.
